feat(chat): underline tab strip for the chat drawer header
The drawer's Tools / Actions strip filled its active segment with the theme
accent across the full panel width, so the switcher outweighed every card it
was introducing, and the two halves read as different kinds of control rather
than a pair. The full-width track also started at y=14, right where the panel's
drag handle sits, which made the handle look like a seam across it.

- Add `GlazeTabBarStyle`: `pill` (unchanged, still the app default for the
  other 11 strips) and `underline` — no container, an accent rule under the
  active tab, tabs sized to their labels
- Split the two layouts into `glaze_tab_bar_strips.dart`; `GlazeTabBar` is now
  a thin dispatcher, and the pill strip reuses `SwipeTabSwitcher` instead of
  repeating its drag maths
- Underline tabs shrink and ellipsize under pressure rather than scrolling, so
  the strip never competes with the swipe-to-switch gesture, and one tween
  drives the label, icon and rule so a tab settles as one object
- Use it in the drawer header, with the top inset moved to 20 so the drag
  handle keeps its own band
- Edit toggle is a bare glyph at rest; the ring now appears only while editing,
  so it marks the state instead of decorating both
- Extract `kDrawerContentTopInset`, shared by both tab bodies, so the header
  height and the content inset cannot drift apart

Verified with `flutter analyze` (no new issues) and `flutter test` (3697
passing). Layout checked by rendering the real widget at 411 dp in EN and RU,
in both tab states and in edit mode.
