+++ docs/UI_KIT.md
@@ -26,7 +26,7 @@ styling a Material widget with `Colors.white.withValues(alpha: 0.05)` and
 |---|---|---|
 | `Scaffold` + `AppBar` | `GlazeScaffold` (`glaze_scaffold.dart`) | Floating glass header for screens **outside** the shell. Screens inside a shell branch pass `useShellHeader: true` + `headerBranchIndex` and publish into the shell's persistent header instead. |
 | `showModalBottomSheet` with a hand-rolled body | `SheetView` (`sheet_view.dart`) or `GlazeBottomSheet.show` (`glaze_bottom_sheet.dart`) | See "Which sheet" below. |
-| `TabBar` / `TabBarView` / `DefaultTabController` | `GlazeTabBar` + `SwipeTabSwitcher` + `TabSlideSwitcher` | The segmented control. `GlazeTabBar` handles tap, swipe-on-the-strip and scrolling past ~2.35 tabs; `SwipeTabSwitcher` adds body swipe; `TabSlideSwitcher` animates the body in the swipe's direction. |
+| `TabBar` / `TabBarView` / `DefaultTabController` | `GlazeTabBar` + `SwipeTabSwitcher` + `TabSlideSwitcher` | The segmented control. `GlazeTabBar` handles tap, swipe-on-the-strip and scrolling past ~2.35 tabs; `SwipeTabSwitcher` adds body swipe; `TabSlideSwitcher` animates the body in the swipe's direction. `style:` picks the look — `pill` (default) for a strip *inside* a screen or sheet, `underline` for one that *heads* a surface, where a filled accent segment would outweigh the content below it (the chat drawer). |
 | `Card` / `Container` with a translucent fill | `GlassSurface` (`glass_surface.dart`) | Reads the active theme preset. Give it `onTap` for a `GlowInkWell` ripple, or `enableRipple: true` for the hover glow without a tap handler. |
 | `ListTile` groups, settings rows | `MenuGroup` + `MenuItem` / `MenuSwitchItem` / `MenuSelectorItem` / `MenuFieldItem` / `MenuRangeItem` / `MenuScriptItem` / `MenuSubHeader` / `MenuCollapsibleSection` (`menu_group.dart`) | The standard settings/form row set. `MenuGroup` already supplies the surface, the border and the 16 px horizontal margin. |
 | `ElevatedButton` / `OutlinedButton` / `FilledButton` in a toolbar | `GlassSurface` tile with `onTap`, or `GlazePillButton` (`glaze_scaffold.dart`) | There is no generic Glaze button — a tile built on `GlassSurface` *is* the button. |
