+++ lib/features/chat/widgets/quick_replies_panel.dart
@@ -142,7 +142,7 @@ class _QuickRepliesPanelState extends ConsumerState<QuickRepliesPanel> {
 
     final content = RawScrollbar(
       controller: _scrollController,
-      padding: const EdgeInsets.only(top: 60),
+      padding: const EdgeInsets.only(top: kDrawerContentTopInset),
       thickness: 3,
       radius: const Radius.circular(3),
       thumbColor: Colors.white24,
@@ -155,7 +155,7 @@ class _QuickRepliesPanelState extends ConsumerState<QuickRepliesPanel> {
               controller: _scrollController,
               padding: EdgeInsets.fromLTRB(
                 12,
-                60,
+                kDrawerContentTopInset,
                 12,
                 16 + MediaQuery.of(context).padding.bottom,
               ),
